+++ plan/plan.go
@@ -21,7 +21,6 @@ import (
 	"github.com/juju/errors"
 	"github.com/pingcap/tidb/context"
 	"github.com/pingcap/tidb/expression"
-	"github.com/pingcap/tidb/plan/util"
 	"github.com/pingcap/tidb/util/codec"
 	"github.com/pingcap/tidb/util/types"
 	"github.com/pingcap/tipb/go-tipb"
@@ -440,7 +439,7 @@ func (p *basePlan) ReplaceParent(parent, newPar Plan) error {
 			return nil
 		}
 	}
-	return util.SystemInternalErrorType.Gen("ReplaceParent Failed!")
+	return SystemInternalErrorType.Gen("ReplaceParent Failed!")
 }
 
 // ReplaceChild means replace a child with another one.
@@ -451,7 +450,7 @@ func (p *basePlan) ReplaceChild(child, newChild Plan) error {
 			return nil
 		}
 	}
-	return util.SystemInternalErrorType.Gen("ReplaceChildren Failed!")
+	return SystemInternalErrorType.Gen("ReplaceChildren Failed!")
 }
 
 // Parents implements Plan Parents interface.
