+++ plan/validator.go
@@ -22,7 +22,6 @@ import (
 	"github.com/pingcap/tidb/infoschema"
 	"github.com/pingcap/tidb/mysql"
 	"github.com/pingcap/tidb/parser"
-	"github.com/pingcap/tidb/plan/util"
 	"github.com/pingcap/tidb/util/types"
 )
 
@@ -275,7 +274,7 @@ func (v *validator) checkAlterTableGrammar(stmt *ast.AlterTableStmt) {
 		case ast.AlterTableOption:
 			for _, opt := range spec.Options {
 				if opt.Tp == ast.TableOptionAutoIncrement {
-					v.err = util.ErrAlterAutoID
+					v.err = ErrAlterAutoID
 					return
 				}
 			}
